requirements.txt
openai
openai-whisper
python-dotenv
azure-cognitiveservices-speech